requirements.txt
# Booth
python-escpos
gpiozero
RPi.GPIO
lgpio
pyusb
picamera2
numpy
Pillow
evdev

# Manager
flask
